Split ValueResolverBackend
