+++ it/java8/src/main/java/online/sharedtype/it/java8/TempClass.java
@@ -1,9 +1,8 @@
 package online.sharedtype.it.java8;
 
+import lombok.RequiredArgsConstructor;
 import online.sharedtype.SharedType;
-import online.sharedtype.it.java8.other.OtherConstants;
 
-@SharedType(constantNamespaced = SharedType.OptionalBool.FALSE, includes = SharedType.ComponentType.CONSTANTS)
 public class TempClass extends IgnoredSuperClassB {
