+++ processor/src/main/java/online/sharedtype/processor/parser/value/ValueResolveContext.java
@@ -4,8 +4,11 @@ import com.sun.source.tree.BlockTree;
 import com.sun.source.tree.Scope;
 import com.sun.source.tree.Tree;
 import com.sun.source.util.Trees;
-import lombok.Builder;
+import lombok.EqualsAndHashCode;
 import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+import lombok.ToString;
+import online.sharedtype.processor.context.Context;
 import online.sharedtype.processor.support.exception.SharedTypeException;
 
 import javax.lang.model.element.Element;
@@ -14,21 +17,34 @@ import javax.lang.model.element.TypeElement;
 import javax.lang.model.util.Elements;
 import javax.lang.model.util.Types;
 
+@EqualsAndHashCode(of = {"fieldElement", "tree", "enclosingTypeElement"})
 @Getter
-@Builder(toBuilder = true)
 final class ValueResolveContext {
     // utils:
+    private final Context ctx;
     private final Trees trees;
     private final Elements elements;
     private final Types types;
 
     // original values:
     private final Element fieldElement;
-    private final TypeElement ctxTypeElement;
 
     // current values:
     private final Tree tree;
     private final TypeElement enclosingTypeElement;
+
+    ValueResolveContext(Context ctx, Element fieldElement, Tree tree, TypeElement enclosingTypeElement) {
+        this.ctx = ctx;
+        this.trees = ctx.getTrees();
+        this.elements = ctx.getProcessingEnv().getElementUtils();
+        this.types = ctx.getProcessingEnv().getTypeUtils();
+        this.fieldElement = fieldElement;
+        this.tree = tree;
+        this.enclosingTypeElement = enclosingTypeElement;
+    }
+    public static Builder builder(Context ctx) {
+        return new Builder(ctx);
+    }
 
     Scope getScope() {
         return trees.getScope(trees.getPath(enclosingTypeElement));
@@ -48,5 +64,34 @@ final class ValueResolveContext {
 
     PackageElement getPackageElement() {
         return elements.getPackageOf(enclosingTypeElement);
+    }
+
+    public Builder toBuilder() {
+        return new Builder(this.ctx).fieldElement(this.fieldElement).tree(this.tree).enclosingTypeElement(this.enclosingTypeElement);
+    }
+
+    @ToString
+    @RequiredArgsConstructor
+    static final class Builder {
+        private final Context ctx;
+        private Element fieldElement;
+        private Tree tree;
+        private TypeElement enclosingTypeElement;
+
+        public Builder fieldElement(Element fieldElement) {
+            this.fieldElement = fieldElement;
+            return this;
+        }
+        public Builder tree(Tree tree) {
+            this.tree = tree;
+            return this;
+        }
+        public Builder enclosingTypeElement(TypeElement enclosingTypeElement) {
+            this.enclosingTypeElement = enclosingTypeElement;
+            return this;
+        }
+        public ValueResolveContext build() {
+            return new ValueResolveContext(this.ctx, this.fieldElement, this.tree, this.enclosingTypeElement);
+        }
     }
 }
