+++ processor/src/main/java/online/sharedtype/processor/parser/value/ValueResolverBackend.java
@@ -1,0 +1,9 @@
+package online.sharedtype.processor.parser.value;
+
+interface ValueResolverBackend {
+    Object recursivelyResolve(ValueResolveContext parsingContext);
+
+    static ValueResolverBackend create(ValueResolver valueResolver) {
+        return new ValueResolverBackendImpl(valueResolver);
+    }
+}
