+++ processor/src/test/java/online/sharedtype/processor/parser/value/EnumValueResolverTest.java
@@ -1,10 +1,15 @@
 package online.sharedtype.processor.parser.value;
 
+import com.sun.source.tree.LiteralTree;
 import online.sharedtype.SharedType;
 import online.sharedtype.processor.context.ContextMocks;
 import online.sharedtype.processor.context.TypeElementMock;
+import online.sharedtype.processor.domain.Constants;
+import online.sharedtype.processor.domain.type.ConcreteTypeInfo;
 import online.sharedtype.processor.domain.value.EnumConstantValue;
 import online.sharedtype.processor.parser.type.TypeInfoParser;
+import online.sharedtype.processor.support.exception.SharedTypeException;
+import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.mockito.ArgumentCaptor;
 
@@ -15,27 +20,57 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.eq;
 import static org.mockito.Mockito.mock;
+import static org.mockito.Mockito.reset;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
 final class EnumValueResolverTest {
     private final ContextMocks ctxMocks = new ContextMocks();
     private final TypeInfoParser typeInfoParser = mock(TypeInfoParser.class);
-    private final EnumValueResolver resolver = new EnumValueResolver(ctxMocks.getContext(), typeInfoParser);
+    private final ValueResolverBackend valueResolverBackend = mock(ValueResolverBackend.class);
+    private final EnumValueResolver resolver = new EnumValueResolver(ctxMocks.getContext(), typeInfoParser, valueResolverBackend);
 
     private final ArgumentCaptor<String> msgCaptor = ArgumentCaptor.forClass(String.class);
     private final TypeElementMock enumType = ctxMocks.typeElement("com.github.cuzfrog.EnumA");
+
+    @BeforeEach
+    void setup() {
+        when(valueResolverBackend.recursivelyResolve(any())).then(invoc -> {
+            var tree = invoc.getArgument(0, ValueResolveContext.class).getTree();
+            if (tree instanceof LiteralTree) {
+                return ((LiteralTree) tree).getValue();
+            }
+            throw new AssertionError("Failed test. Unexpected tree: " + tree);
+        });
+    }
+
+    @Test
+    void simpleEnum() {
+        var enumConstant1 = ctxMocks.declaredTypeVariable("VALUE1", enumType.type())
+            .withElementKind(ElementKind.ENUM_CONSTANT)
+            .element();
+        enumType.withEnclosedElements(enumConstant1);
+        var enumTypeInfo = ConcreteTypeInfo.builder().qualifiedName("com.github.cuzfrog.EnumA").kind(ConcreteTypeInfo.Kind.ENUM).build();
+        when(typeInfoParser.parse(enumType.type(), enumType.element())).thenReturn(enumTypeInfo);
+
+        EnumConstantValue value1 = (EnumConstantValue) resolver.resolve(enumConstant1, enumType.element());
+        assertThat(value1.getValue()).isEqualTo("VALUE1");
+        assertThat(value1.getEnumConstantName()).isEqualTo("VALUE1");
+        assertThat(value1.getValueType()).isEqualTo(enumTypeInfo);
+    }
 
     @Test
     void enumValueMarkedOnField() {
         var field1 = ctxMocks.primitiveVariable("field1", TypeKind.INT);
         var field2 = ctxMocks.primitiveVariable("field2", TypeKind.CHAR).withAnnotation(SharedType.EnumValue.class);
+        var valueTree1 = ctxMocks.literalTree('a');
+        var valueTree2 = ctxMocks.literalTree('b');
         var enumConstant1 = ctxMocks.declaredTypeVariable("Value1", enumType.type())
             .withElementKind(ElementKind.ENUM_CONSTANT)
             .ofTree(
                 ctxMocks.variableTree().withInitializer(
                     ctxMocks.newClassTree().withArguments(
-                        ctxMocks.literalTree(100), ctxMocks.literalTree('a')
+                        ctxMocks.literalTree(100), valueTree1
                     ).getTree()
                 )
             )
@@ -45,7 +80,7 @@ final class EnumValueResolverTest {
             .ofTree(
                 ctxMocks.variableTree().withInitializer(
                     ctxMocks.newClassTree().withArguments(
-                        ctxMocks.literalTree(200), ctxMocks.literalTree('b')
+                        ctxMocks.literalTree(200), valueTree2
                     ).getTree()
                 )
             )
@@ -57,10 +92,12 @@ final class EnumValueResolverTest {
             field2.element()
         );
         when(ctxMocks.getContext().isAnnotatedAsEnumValue(field2.element())).thenReturn(true);
+        when(typeInfoParser.parse(field2.type(), enumType.element())).thenReturn(Constants.CHAR_TYPE_INFO);
 
         EnumConstantValue value1 = (EnumConstantValue) resolver.resolve(enumConstant1, enumType.element());
         assertThat(value1.getEnumConstantName()).isEqualTo(enumConstant1.getSimpleName().toString());
         assertThat(value1.getValue()).isEqualTo('a');
+        assertThat(value1.getValueType()).isEqualTo(Constants.CHAR_TYPE_INFO);
         var value2 = resolver.resolve(enumConstant2, enumType.element());
         assertThat(value2.getValue()).isEqualTo('b');
     }
@@ -91,7 +128,7 @@ final class EnumValueResolverTest {
     }
 
     @Test
-    void failWhenEnumValueTypeIsNotLiteral() {
+    void logErrorWhenBackendThrowsException() {
         var classB = ctxMocks.typeElement("com.github.cuzfrog.EnumB").withElementKind(ElementKind.CLASS);
         var field1 = ctxMocks.declaredTypeVariable("field1", classB.type())
             .withElementKind(ElementKind.FIELD)
@@ -110,8 +147,11 @@ final class EnumValueResolverTest {
         enumType.withEnclosedElements(enumConstant1, field1);
         when(ctxMocks.getContext().isAnnotatedAsEnumValue(field1)).thenReturn(true);
 
+        reset(valueResolverBackend);
+        when(valueResolverBackend.recursivelyResolve(any())).thenThrow(new SharedTypeException("test msg"));
+
         resolver.resolve(enumConstant1, enumType.element());
         verify(ctxMocks.getContext()).error(eq(enumType.element()), msgCaptor.capture(), any(Object[].class));
-        assertThat(msgCaptor.getValue()).contains("Only literals are supported");
+        assertThat(msgCaptor.getValue()).contains("Failed to parse argument value");
     }
 }
